+++ flumine/streams/historicalstream.py
@@ -10,34 +10,32 @@ from ..event.event import MarketBookEvent
 logger = logging.getLogger(__name__)
 
 
-class MarketStream(BaseStream):
+class HistoricalStream(BaseStream):
+
+    MAX_LATENCY = 1e100
+
     @retry(wait=wait_exponential(multiplier=1, min=2, max=20))
     def run(self) -> None:
-        logger.info("Starting MarketStream")
+        logger.info("Starting HistoricalStream")
+        self.stream_id = 0  # default for historical data
 
         if not self._output_thread.is_alive():
             logger.info("Starting output_thread {0}".format(self._output_thread))
             self._output_thread.start()
 
-        self._stream = self.trading.streaming.create_stream(
-            unique_id=self.stream_id, listener=self._listener
-        )
-        try:
-            self.stream_id = self._stream.subscribe_to_markets(
-                market_filter=self.market_filter,
-                market_data_filter=self.market_data_filter,
-                conflate_ms=self.conflate_ms,
-                initial_clk=self._listener.initial_clk,  # supplying these two values allows a reconnect
-                clk=self._listener.clk,
+        for market in self.market_filter["markets"]:
+            self._stream = self.trading.streaming.create_historical_stream(
+                directory=market, listener=self._listener
             )
-            self._stream.start()
-        except BetfairError:
-            logger.error("MarketStream run error", exc_info=True)
-            raise
-        except Exception:
-            logger.critical("MarketStream run error", exc_info=True)
-            raise
-        logger.info("Stopped MarketStream {0}".format(self.stream_id))
+            try:
+                self._stream.start()
+            except BetfairError:
+                logger.error("HistoricalStream run error", exc_info=True)
+                raise
+            except Exception:
+                logger.critical("HistoricalStream run error", exc_info=True)
+                raise
+        logger.info("Stopped HistoricalStream {0}".format(self.stream_id))
 
     def handle_output(self) -> None:
         """Handles output from stream.
@@ -53,4 +51,6 @@ class MarketStream(BaseStream):
                 )
             self.flumine.handler_queue.put(MarketBookEvent(market_books))
 
-        logger.info("Stopped output_thread (MarketStream {0})".format(self.stream_id))
+        logger.info(
+            "Stopped output_thread (HistoricalStream {0})".format(self.stream_id)
+        )